+++ next-env.d.ts
@@ -9,5 +9,6 @@ declare module '*.styl' {
 declare module '@deck.gl/core'
 declare module '@deck.gl/react'
 declare module '@deck.gl/layers'
+declare module '@luma.gl/core'
 
 declare var __CLIENT__: boolean
